+++ src/udtBleConnection.ts
@@ -150,7 +150,7 @@ export class UdtBleConnection {
         } catch (error) {
             this.logger.error(`Tower Connection Error: ${error}`, '[UDT][BLE]');
             this.isConnected = false;
-            this.callbacks.onTowerDisconnect();
+            throw error;
         }
     }